* master: (642 commits)
  Add the Vanilla namespace to composer’s autoloader (#3739)
  Add caching to UserModel::getRoles (#3723)
  Empty strings should return null test
  Tweak dbencode/dbdecode
  Don't assume zero index in Gdn_SQLDriver::where
  Change error handling on dbencode/dbdecode
  Fix bug with UserModel->definePermissions and caching
  Fix scrutinizer reference assignment spacing.
  Rollback overzealous language attributes
  Refactor database serialization calls (#3734)
  Update category checking in DiscussionModel::getAnnouncements
  Fire "CategoryWatch" as an event of CategoryModel (#3740)
  Remove unused and undefined variables from mergeActivities method.
  Revert "Remove unused and undefined variables from mergeActivities method."
  Remove unused and undefined variables from mergeActivities method.
  Tweak PHPUnit config
  Add and update lang attributes on HTML tags (#3704)
  Remove @error suppression in a few places (#3721)
  Remove hidden locales from settings listings unless they are enabled (#3732)
  Upgrade htmLawed to v2.1.22 (#3727)
  ...

+++ applications/dashboard/controllers/class.logcontroller.php
@@ -2,7 +2,7 @@
 /**
  * Non-activity action logging.
  *
- * @copyright 2009-2015 Vanilla Forums Inc.
+ * @copyright 2009-2016 Vanilla Forums Inc.
  * @license http://www.opensource.org/licenses/gpl-2.0.php GNU GPL v2
  * @package Dashboard
  * @since 2.0
@@ -31,11 +31,16 @@ class LogController extends DashboardController {
      * @param string $Action Type of action.
      * @param array $LogIDs Numeric IDs of items to confirm.
      */
-    public function confirm($Action, $LogIDs = '') {
+    public function confirm() {
+        if (!Gdn::request()->isAuthenticatedPostBack(true)) {
+            throw new Exception('Requires POST', 405);
+        }
         $this->permission(array('Garden.Moderation.Manage', 'Moderation.Spam.Manage', 'Moderation.ModerationQueue.Manage'), false);
+        $Action = Gdn::request()->post('Action', false);
+        $LogIDs = Gdn::request()->post('IDs', false);
 
-        $this->Form->InputPrefix = '';
-        $this->Form->IDPrefix = 'Confirm_';
+        $this->Form->addHidden('LogIDs', $LogIDs);
+       $this->Form->IDPrefix = 'Confirm_';
 
         if (trim($LogIDs)) {
             $LogIDArray = explode(',', $LogIDs);
@@ -57,9 +62,8 @@ class LogController extends DashboardController {
         $this->setData('Users', $Users);
 
         $this->setData('Action', $Action);
-        $this->setData('ActionUrl', url("/log/$Action?logids=".urlencode($LogIDs)));
+        $this->setData('ActionUrl', url('/log/'.$Action));
         $this->setData('ItemCount', count($LogIDArray));
-
         $this->render();
     }
 
@@ -101,9 +105,12 @@ class LogController extends DashboardController {
      *
      * @param array $LogIDs Numeric IDs of logs to delete.
      */
-    public function delete($LogIDs) {
+    public function delete() {
+        if (!Gdn::request()->isAuthenticatedPostBack(true)) {
+            throw new Exception('Requires POST', 405);
+        }
         $this->permission(array('Garden.Moderation.Manage', 'Moderation.ModerationQueue.Manage'), false);
-        // Grab the logs.
+        $LogIDs = Gdn::request()->post('LogIDs');
         $this->LogModel->delete($LogIDs);
         $this->render('Blank', 'Utility');
     }
@@ -112,13 +119,12 @@ class LogController extends DashboardController {
      * Delete spam and optionally delete the users.
      * @param type $LogIDs
      */
-    public function deleteSpam($LogIDs) {
-        $this->permission(array('Garden.Moderation.Manage', 'Moderation.Spam.Manage'), false);
-
-        if (!$this->Request->isPostBack()) {
-            throw permissionException('Javascript');
+    public function deleteSpam() {
+        if (!Gdn::request()->isAuthenticatedPostBack(true)) {
+            throw new Exception('Requires POST', 405);
         }
-
+        $this->permission(array('Garden.Moderation.Manage', 'Moderation.Spam.Manage'), false);
+        $LogIDs = Gdn::request()->post('LogIDs');
         $LogIDs = explode(',', $LogIDs);
 
         // Ban the appropriate users.
@@ -255,7 +261,7 @@ class LogController extends DashboardController {
         Gdn_Theme::section('Dashboard');
         $this->addJsFile('log.js');
         $this->addJsFile('jquery.expander.js');
-        $this->Form->InputPrefix = '';
+        $this->addJsFile('jquery-ui.js');
     }
 
     /**
@@ -306,8 +312,12 @@ class LogController extends DashboardController {
      *
      * @param array $LogIDs List of log IDs.
      */
-    public function restore($LogIDs) {
+    public function restore() {
+        if (!Gdn::request()->isAuthenticatedPostBack(true)) {
+            throw new Exception('Requires POST', 405);
+        }
         $this->permission(array('Garden.Moderation.Manage', 'Moderation.Spam.Manage', 'Moderation.ModerationQueue.Manage'), false);
+        $LogIDs = Gdn::request()->post('LogIDs');
 
         // Grab the logs.
         $Logs = $this->LogModel->getIDs($LogIDs);
@@ -322,12 +332,12 @@ class LogController extends DashboardController {
         $this->render('Blank', 'Utility');
     }
 
-    public function notSpam($LogIDs) {
-        $this->permission(array('Garden.Moderation.Manage', 'Moderation.Spam.Manage'), false);
-
-        if (!$this->Request->isPostBack()) {
-            throw permissionException('Javascript');
+    public function notSpam() {
+        if (!Gdn::request()->isAuthenticatedPostBack(true)) {
+            throw new Exception('Requires POST', 405);
         }
+        $this->permission(array('Garden.Moderation.Manage', 'Moderation.Spam.Manage'), false);
+        $LogIDs = Gdn::request()->post('LogIDs');
 
         $Logs = array();
